+++ python/examples/rigid_cubes/main.py
@@ -3,8 +3,8 @@ import dartpy as dart
 
 
 class MyWorldNode(dart.gui.osg.RealTimeWorldNode):
-    def __init__(self, world):
-        super(MyWorldNode, self).__init__(world)
+    def __init__(self, world, shadow):
+        super(MyWorldNode, self).__init__(world, shadow)
 
     def customPreStep(self):
         pass
@@ -14,9 +14,11 @@ def main():
     world = dart.utils.SkelParser.readWorld("dart://sample/skel/cubes.skel")
     world.setGravity([0, -9.81, 0])
 
-    node = MyWorldNode(world)
 
     viewer = dart.gui.osg.Viewer()
+    shadow = dart.gui.osg.WorldNode.createDefaultShadowTechnique(viewer)
+
+    node = MyWorldNode(world, shadow)
     viewer.addWorldNode(node)
 
     viewer.setUpViewInWindow(0, 0, 640, 480)